third_party/WebKit/LayoutTests/fast/css-grid-layout/grid-item-start-before-get-set-expected.txt

Test that setting and getting grid-start and grid-before works as expected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test getting -webkit-grid-start and -webkit-grid-before set through CSS
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-start') is 'auto'
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-column') is 'auto / auto'
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-before') is 'auto'
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-row') is 'auto / auto'
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') is gridAfterValue
FAIL getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') should be span 1 first / auto. Was auto / auto.
FAIL getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') should be span 1 first. Was auto.
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') is gridEndValue
FAIL getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') should be span 1 last / auto. Was auto / auto.
FAIL getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') should be span 1 last. Was auto.
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') is gridAfterValue
FAIL getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') should be span 2 first / auto. Was auto / auto.
FAIL getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') should be span 2 first. Was auto.
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') is gridEndValue
FAIL getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') should be span 3 last / auto. Was auto / auto.
FAIL getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') should be span 3 last. Was auto.
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') is gridAfterValue

Test the initial value
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-start') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-column') is 'auto / auto'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-before') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-row') is 'auto / auto'

Test getting and setting grid-start and grid-before through JS
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') is gridAfterValue
FAIL getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') should be span 1 nav / auto. Was auto / auto.
FAIL getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') should be span 1 nav. Was auto.
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') is gridEndValue
FAIL getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') should be span 1 last / auto. Was auto / auto.
FAIL getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') should be span 1 last. Was auto.
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') is gridAfterValue

Test setting grid-start and grid-before to 'inherit' through JS
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') is gridAfterValue

Test setting grid-start and grid-before to 'initial' through JS
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') is gridAfterValue

Test setting grid-start and grid-before back to 'auto' through JS
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-start') is '18'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-column') is '18 / auto'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-before') is '66'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-row') is '66 / auto'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-start') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-column') is 'auto / auto'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-before') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-row') is 'auto / auto'
PASS successfullyParsed is true

TEST COMPLETE
